% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypress.R
\name{has_keypress_support}
\alias{has_keypress_support}
\title{Check if the current platform/terminal supports reading
single keys.}
\usage{
has_keypress_support()
}
\value{
Whether there is support for waiting for individual
keypressses.
}
\description{
Keypress supports terminals (Windows or Unix), including the
RStudio terminal. Other environments, e.g. the macOS R.app, the
Windowd Rgui, Emacs ESS, etc. are not supported.
}
\examples{
has_keypress_support()
}
\seealso{
Other keypress: 
\code{\link{keypress}()}
}
\concept{keypress}
