% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypress.R
\name{restore_term_status}
\alias{restore_term_status}
\title{Restore previously saved terminal state}
\usage{
restore_term_status()
}
\description{
For Linux/Unix and OSX terminals, restore the current flags relating to
behaviour of the terminal used to call the R script.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
For Linux/Unix and OSX terminals using non-blocking keypress, special
settings need applying to prevent characters being echoed to the terminal
as well as being interpreted by keypress. However, any changes made
to the settings will by default persist after exiting your R script.
Therefore, should you need non-blocking keypresses from one of these
terminals, you should save the terminal settings and turn off echo near
the beginning of your code, and then restore the original terminal
settings on exit.

This will be safely ignored on Windows, as it is not needed.
}
\seealso{
Other terminal functions: 
\code{\link{save_term_status}()},
\code{\link{set_term_echo}()}
}
\concept{terminal functions}
