% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypress.R
\name{set_term_echo}
\alias{set_term_echo}
\title{Set echo flag of terminal.}
\usage{
set_term_echo(echo = FALSE)
}
\description{
For Linux/Unix and OSX terminals, set the echo flag to ON or OFF.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
You will need to set echo to FALSE if you want the R script to handle all
of the keypress-related behaviour, without any keys being echoed in the
terminal by the operating system, when using non-blocking keypress. This is
not necessary when running in a Windows command prompt, and will be ignored.
}
\examples{
set_term_echo(FALSE)
}
\seealso{
Other terminal functions: 
\code{\link{restore_term_status}()},
\code{\link{save_term_status}()}
}
\concept{terminal functions}
