% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypress.R
\name{with_no_echo}
\alias{with_no_echo}
\title{Call a function with echo suppressed}
\usage{
with_no_echo(func)
}
\arguments{
\item{func}{The function to be run with echo suppressed}
}
\description{
For Linux/Unix and OSX terminals, suppress key echoes from the terminal.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
You will need to set echo to FALSE if you want the R script to handle all
of the keypress-related behaviour, without any keys being echoed in the
terminal by the operating system, when using non-blocking keypress. This is
not necessary when running in a Windows command prompt, and will be safely
ignored.
}
\concept{terminal functions}
