% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keypress.R
\name{has_keypress_support}
\alias{has_keypress_support}
\title{Check if the current platform/terminal supports reading
single keys.}
\usage{
has_keypress_support()
}
\value{
Whether there is support for waiting for individual
keypressses.
}
\description{
Check if the current platform/terminal supports reading
single keys.
}
\examples{
has_keypress_support()
}
\seealso{
Other keypress: \code{\link{keypress}}
}

