% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypress.R
\name{keypress}
\alias{keypress}
\title{Read a single keypress at the terminal}
\usage{
keypress(block = TRUE)
}
\arguments{
\item{block}{Whether to wait for a key press, if there is none
available now.}
}
\value{
The key pressed, a character scalar. For non-blocking reads
  NA is returned if no keys are available.
}
\description{
It currently only works at Linux/Unix and OSX terminals,
and at the Windows command line. see \code{\link{has_keypress_support}}.
}
\details{

}
\examples{
\dontrun{
x <- keypress()
cat("You pressed key", x, "\n")
}
}
\seealso{
Other keypress: 
\code{\link{has_keypress_support}()}
}
\concept{keypress}
