% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keypress.R
\name{keypress}
\alias{keypress}
\title{Wait for a single keypress at the terminal}
\usage{
keypress()
}
\value{
The key pressed, a character scalar.
}
\description{
It currently only works at Linux/Unix and OSX terminals,
see \code{\link{has_keypress_support}}.
}
\details{
The following special keys are supported:
\itemize{
  \item Arrow keys: \sQuote{up}, \sQuote{down}, \sQuote{right},
    \sQuote{left}.
  \item Function keys: from \sQuote{f1} to \sQuote{f12}.
  \item Others: sQuote{home}, \sQuote{end},
    \sQuote{insert}, \sQuote{delete}, \sQuote{pageup},
    \sQuote{pagedown}.
}
}
\examples{
\dontrun{
x <- keypress()
cat("You pressed key", x, "\\n")
}
}
\seealso{
Other keypress: \code{\link{has_keypress_support}}
}

